Split new/legacy tests
