+++ cross/apple/Tests/PartoutTests/CollectionExtensionsTests.swift
@@ -2,6 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 
+@testable import PartoutCore
 import Testing
 
 struct CollectionExtensionsTests {
