+++ cross/apple/Tests/PartoutTests/DataManipulationTests.swift
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 
-import PartoutLegacyCore
+import PartoutCore
 import Testing
 
 struct DataManipulationTests {
