+++ cross/apple/Tests/PartoutTests/StaticKeyTests.swift
@@ -2,8 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 
-import PartoutLegacyCore
-@testable import PartoutOpenVPN
+@testable import PartoutCore
 import Testing
 
 struct StaticKeyTests {
