+++ cross/apple/Tests/PartoutTests/TaggedModuleTests.swift
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import Foundation
-import PartoutLegacyCore
+import PartoutCore
 import Testing
 
 struct TaggedModuleTests {
